refactor to clean up types
- Fixes various Typescript errors
- Adds APIs to better encapsulate shadow DOM features
